DEF[#79403]: 温控器 modbus 协议完善

See merge request iotcenter/IotCenter!1688

+++ iot-center-accessmeter/src/main/java/com/yjh/accessmeter/module/device/entity/IotDeviceData.java
@@ -58,5 +58,5 @@ public class IotDeviceData {
     /**
      * 系数
      */
-    private Integer magnificationCoefficient;
+    private Float magnificationCoefficient;
 }
